Change type on CSV to match for export

+++ src/SFA.DAS.Courses.Domain/ImportTypes/ApprenticeshipFundingCsv.cs
@@ -7,6 +7,6 @@ namespace SFA.DAS.Courses.Domain.ImportTypes
         public int ApprenticeshipCode { get; set; }
         public DateTime EffectiveFrom { get; set; }
         public DateTime? EffectiveTo { get; set; }
-        public int MaxEmployerLevyCap { get; set; }
+        public decimal MaxEmployerLevyCap { get; set; }
     }
 }
